Update mikes_initial_notes.txt
Realised we had one extra item missing from our discussion yesterday and that was your new run command/search. I think it'd be good just to stay on top of this file as a primitive checklist so if we feel like just adding something we can consult the list and try to avoid feature creep. Made the entries a bit more explicit about what the feature is and why it's needed and then tried to priorities them, not in order of easiest to build but most important for fleshing out the foundations of this beast.